requirements.txt
langchain
openai
faiss-cpu
pypdf
tiktoken
streamlit
python-dotenv